refactor(test): add missing test, split testcase
